schem: better usage of shift/storage register clock lines
Now SPI_CLK connects to (595's) internal shift register clock
line, and data on SPI_MOSI is always shifted into this register.
However, the contents are only transmitted to storage register and
external output on positive edge of SPI_SLAVE2 slave selection line.
This is NOT the general /SS of SPI, which is actually not used.
SPI_SLAVE2 is currently on pin PC1 of atmega168.

Previous setup of common clock for both registers didn't seem to
work on breadboard. Parallel outputs would show intermediary
stages of internal shift register, while the data was still being
shifted through the serial input.