code: experimental counter usage to determine second zero crossing
NOTE: Current zero crossing schematic only allows detecting one AC
half-cycle. Depending on where the current-limiting resistor on the
output of the optocouple is placed, you can have a defined rising or
falling edge, but not both. When the transistor is non-conducting, it
seems to act as a capacitor.

This introduces the need to use a counter to detect every
other zero crossing. It is annoying and it complicates the code.

A better ZC detector is required, where both rising and falling edge
are usable.

On the other side, a counter is needed anyway to determine the firing
angle. At 16Mhz CPU clock and 50 Hz mains frequency (100 zc per second),
there are

16 000 000 / 100 = 160 000 instructions between zero crossings
160 000 / 256    =     625 instructions between eq. spaced firing angles

With the counter prescaled to F_CPU/256, there are

16 000 000 / 256 =  62 500 counter ticks per second
62 500 / 100     =     625 counter ticks between zero crossings