schem: finished zc/dimming module, added digipot symbol, but see details
Using a digipot seems to be impossible now. Added complexity aside, they
have a limit on maximum current in the range from 0.5 mA to 20 mA (and
the latter are expensive). Charging the 555 cap requires currents of
VCC/100 (maximum value, when pot resistance is 0). Even for 5 V, this is
50 mA.

So scrap this digipot wankery.

Perhaps a dedicated microcontroller might be up for the task. If it
chooses its clock crystal wisely, it can detect ZC, receive delay values
from the Master, and issue pulses when necessary.

And you better not ask a hobbit to do it.