code: working zc/dimming, no flicker
Seems that dimmer module's ZC interrupt was sometimes called in the
middle of angle interrupt (INT0 and Timer/Counter0 Output Compare A,
respectively). This caused the control light to turn on for the next 10
ms, even if it didn't have to.

TIMER0_COMPA_vect now has ISR_BLOCK. It is a pretty short interrupt, so
shouldn't interfere with INT0_vect.

ZC duration calibration is extremely simplified, also for debugging
purposes.

The dirty changes in transceiver module are for debugging: they make the
dmx_value rise and fall gradually over a period of approx. 25,5 seconds.

Breadboard-wise, the 47nF cap of the dimmer module has been removed to
seemingly no detriment. EMI is still to be measured and dealt with.