schem: many changes, see full description
- Proper DMX repeating, connected Universe 1 input receiver to output
driver.

- Master connects to USART through 3-pin wire connection. It carries
VCC, RXD and GND. USART TXD and ~RE are not used.

- Instead, OUT_SS2 and OUT_SS3 (Slave Select interrupt lines) are connected
where USART TXD and ~RE were.

- PREHEAT and MAXVAL are connected to PC4 and PC5 of Master, where OUT_SS2
and OUT_SS3 were. These are used by Analog-Digital Converter and determine
the appropriate DMX values. They are connected to two potentiometers on
the front panel.

The ATmega168 datasheet states PC0-PC3 are connected to "analog VCC",
which is probably AVCC, and PC4-PC5 to "digital VCC", which is probably
VCC. The datasheet states if these pins are used as digital outputs it
may interfere with proper analog-to-digital conversion. However, state
change on PC0-PC3 do not affect PC4-PC5, and vice versa. Since PC0-PC5
were all previously used as digital outputs to Slaves, it is probably
wiser to use PC4-PC5.

- More decoupling capacitors - another one for VCC, one for AREF (which
is connected internally to the ADC).

- Also as per datasheet, a small inductive element (10 uH) added to AVCC
"to reduce noise".

- Panel connector is now one 16-pin instead of many (2-, 2-, 5-, etc.).

- As mentioned, two pots on panel to set preheat and maximum value.

- Two 12V connectors for fans for each phase.

- RESET button from panel also connected to Slaves.

- One pin on Slave left unused, so introduce a 3-pin connector for
"custom" extensions.

- Some prettifies.