code:trans: use received preheat and maxval values
This is done when sending the final values to the slaves instead of when
receiving them on USART. The latter is rather more time-dependent.

Right now, the appropriate chanval is modified in memory, so if preheat
and maxval change, the chanval will stay the same as the previous limiting
value. This will change soon.