schem: don't buffer dmx512, pass thru directly
Thought putting a retransmitter for DMX512 data was a good idea.
However, this is not so in real life. If several dimmers are chained
in a rack and not all of them are used, some dimmers may be turned off
(to use less power). In such case, the buffer's SN75176 will prevent
the signal from passing on, since they have no power.

So this now only has one SN75176, the one proper to translate
differential signals to simple USART, to be passed on to the master
uC.

pcb: some leftovers from last commit, which happened half a year ago.