schem: change ATMega168 package from DIP28 to TQFP32 code: minor changes to ADC handling
DIP28 was used with a protoboard. On a prototype board TQFP32 is
to be used (better local availability, smaller footprint).

The latter has four more pins: VCC, GND and two more ADC channels,
ADC6 and ADC7. These can be used for ADC only, and since we're using
two ADC channels for maximum value and preheat settings, it is better
to leave GPIOs PC4 and PC5 unused.

Note that the code currently is incomplete, it doesn't read these
two settings or switch between ADC channels.