src: working multi-channel operation, dirty code
Removed fancy code, which is also hard to read and debug.

The version in this commit has all four channels working. However, there seem to be some stability issues.

First, the slave enters the "new data from Master" interrupt spontaneously on start-up. Or so it seems, since the green debug led comes on.

Second, there is noticeable flicker in dimmed lamp output. The reason for this also unknown. Perhaps it's my test setup. The dimmer module, soldered on protoboard, doesn't have a choke.